requirements.txt
click
pyyaml
requests
loguru
jinja2
pytest